Update AmazonEBS.md
In the last line of the first paragraph EBS volume is referred to in the singular. I think it's best if the singular EBS Volume is used throughout the first paragraph. 
Using the singular makes the writing a little more straightforward and it is somewhat awkward to switch between the plural and the singular.  
I also added a "your" before EC2 instances because that is a AWS Docs style guide type of thing. I kind of prefer the sentence without the "your" but it sounds more like AWS with it in.  I combined the second and third sentence to reduce the redundancy of "An EBS Volume". 
I think the last line in the paragraph is improved with the word "other". Without "other" the "or" implies that a volume can be used like a hard drive or you can create a file system on top of it, but since you can create a filesystem on a hard drive the two actions are not exclusive so the word "other" brings the action of "creating a file system" into the sphere of "using like a hard drive".  
In the second paragraph it makes more sense to use "and" instead of "or". 


Thank you for taking the time to review my pull request.